microservices/interactive-content-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0
google-cloud-pubsub==2.18.4
google-cloud-storage==2.14.0
google-cloud-logging==3.8.0
google-cloud-language==2.11.1
google-cloud-bigquery==3.13.0
aioredis==2.0.1
websockets==11.0.3
python-json-logger==2.0.7
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.23.3
aiohttp==3.9.1
async-timeout==4.0.3
tenacity==8.2.3
prometheus-client==0.19.0
structlog==23.2.0
typing-extensions==4.8.0
python-multipart==0.0.6
email-validator==2.1.0.post1
ujson==5.8.0
orjson==3.9.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
aioboto3==11.3.0
redis==5.0.1
cachetools==5.3.2
dataclasses-json==0.6.3
marshmallow==3.20.1
PyYAML==6.0.1 